moved classes to exceptions module
